Initial IClrInfoProvider implementation
