+++ src/Microsoft.Diagnostics.Runtime.Tests/src/DataTargetTests.cs
@@ -4,6 +4,7 @@
 using System.IO;
 using System.Linq;
 using Microsoft.Diagnostics.Runtime.DacInterface;
+using Microsoft.Diagnostics.Runtime.Implementation;
 using Xunit;
 
 namespace Microsoft.Diagnostics.Runtime.Tests
@@ -20,8 +21,10 @@ namespace Microsoft.Diagnostics.Runtime.Tests
 
             using (ClrRuntime runtime = dt.ClrVersions.Single().CreateRuntime())
             {
-                library = runtime.DacLibrary.OwningLibrary;
-                sosDac = runtime.DacLibrary.SOSDacInterface;
+                ClrRuntimeHelpers runtimeHelpers = (ClrRuntimeHelpers)runtime.DacLibrary;
+
+                library = runtimeHelpers.Library.OwningLibrary;
+                sosDac = runtimeHelpers.SOSDacInterface;
 
                 // Keep library alive
                 library.AddRef();
